refactor: move duplicate RLP code into helper function
